везде показывать телефон агентства в жк

+++ cron/clear.old.news.photos.php
@@ -34,7 +34,7 @@ $db->querys("SET lc_time_names = 'ru_RU';");
 
 // вспомогательные таблицы модуля
 $sys_tables = Config::Get('sys_tables');
-$tables = array('calendar_events');
+$tables = array('articles');
 $sm = 'img/uploads/sm';
 $med = 'img/uploads/med';
 $big = 'img/uploads/big';
@@ -48,18 +48,15 @@ foreach ($tables as $table) {
             " . $sys_tables[$table . '_photos'] . ".name as filename
         FROM " . $sys_tables[$table] . " 
         LEFT JOIN " . $sys_tables[$table . '_photos'] . " ON " . $sys_tables[$table . '_photos'] . ".id = " . $sys_tables[$table] . ".id_main_photo
-        WHERE  " . $sys_tables[$table] . ".date_begin <= '2023-12-31' AND " . $sys_tables[$table . '_photos'] . ".name IS NOT NULL");
+        WHERE  " . $sys_tables[$table] . ".datetime < '2023-03-01' AND " . $sys_tables[$table . '_photos'] . ".name IS NOT NULL");
     foreach ($list as $k => $item) {
         if( !empty( $item['filename'] )) {
             if (file_exists($root . '/' . $sm . '/' . $item['subfolder'] . '/' . $item['filename'])) unlink($root . '/' . $sm . '/' . $item['subfolder'] . '/' . $item['filename']);
-            if (file_exists($root . '/' . $med . '/' . $item['subfolder'] . '/' . $item['filename'])) {
-                echo $item['filename'] . ' - ';
-                unlink($root . '/' . $med . '/' . $item['subfolder'] . '/' . $item['filename']);
-            }
+            if (file_exists($root . '/' . $med . '/' . $item['subfolder'] . '/' . $item['filename'])) unlink($root . '/' . $med . '/' . $item['subfolder'] . '/' . $item['filename']);
             if (file_exists($root . '/' . $big . '/' . $item['subfolder'] . '/' . $item['filename'])) unlink($root . '/' . $big . '/' . $item['subfolder'] . '/' . $item['filename']);
             $db->querys( " UPDATE ".$sys_tables[$table]." SET id_main_photo = 0 WHERE id = ?", $item['id'] );
             $db->querys( " DELETE FROM ".$sys_tables[$table.'_photos']." WHERE id = ?", $item['id_main_photo'] );
-            echo $item['id'].';\n';
+            echo $item['title'].'<br>\n';
         }
     }
